Dev

See merge request bootcamp_myfamily/my_family_flutter!52

+++ lib/features/main/domain/entity/gov_request_response_entity.dart
@@ -3,7 +3,7 @@ import 'package:equatable/equatable.dart';
 
 class GovRequestResponseEntity extends Equatable {
   final String requestId;
-  final int sum;
+  final double sum;
   final String type;
   final DateTime date;
